+++ src/components/multi-autocompleter/multi-autocompleter.test.jsx
@@ -253,4 +253,21 @@ describe('<MultiAutocompleter>', () => {
       expect(await findSelectedOption('test label', 'Bar')).toBeInTheDocument();
     });
   });
+
+  describe('tooltip API', () => {
+    const tooltip = 'I am an input, short and stout.';
+
+    it('applies a description to the input when the help icon is hovered', () => {
+      render(<MultiAutocompleter {...requiredProps} tooltip={tooltip} />);
+      userEvent.hover(screen.getByRole('img', { name: 'More information' }));
+      expect(screen.getByRole('combobox', { name: requiredProps.label })).toHaveDescription(tooltip);
+    });
+
+    it('removes the description to the input when the help icon is unhovered', () => {
+      render(<MultiAutocompleter {...requiredProps} tooltip={tooltip} />);
+      userEvent.hover(screen.getByRole('img', { name: 'More information' }));
+      userEvent.unhover(screen.getByRole('img', { name: 'More information' }));
+      expect(screen.getByRole('combobox', { name: requiredProps.label })).toHaveDescription('');
+    });
+  });
 });