+++ src/components/number/number.test.jsx
@@ -167,6 +167,23 @@ describe('Number', () => {
     });
   });
 
+  describe('tooltip API', () => {
+    const tooltip = 'I am an input, short and stout.';
+
+    it('applies a description to the input when the help icon is hovered', () => {
+      render(<Number {...requiredProps} tooltip={tooltip} />);
+      userEvent.hover(screen.getByRole('img', { name: 'More information' }));
+      expect(screen.getByLabelText(requiredProps.label)).toHaveDescription(tooltip);
+    });
+
+    it('removes the description to the input when the help icon is unhovered', () => {
+      render(<Number {...requiredProps} tooltip={tooltip} />);
+      userEvent.hover(screen.getByRole('img', { name: 'More information' }));
+      userEvent.unhover(screen.getByRole('img', { name: 'More information' }));
+      expect(screen.getByLabelText(requiredProps.label)).toHaveDescription('');
+    });
+  });
+
   describe('value API', () => {
     it('sets the value', () => {
       render(<Number {...requiredProps} value={12} />);